Move VariableNameVisitor to same module as VariableName
